+++ lib/presentation/blocs/category/category_event.dart
@@ -1,0 +1,38 @@
+import 'package:equatable/equatable.dart';
+import 'package:mylist2/data/models/category.dart';
+
+abstract class CategoryEvent extends Equatable {
+  const CategoryEvent();
+
+  @override
+  List<Object?> get props => [];
+}
+
+class LoadCategories extends CategoryEvent {}
+
+class AddCategory extends CategoryEvent {
+  final Category category;
+
+  const AddCategory(this.category);
+
+  @override
+  List<Object?> get props => [category];
+}
+
+class UpdateCategory extends CategoryEvent {
+  final Category category;
+
+  const UpdateCategory(this.category);
+
+  @override
+  List<Object?> get props => [category];
+}
+
+class DeleteCategory extends CategoryEvent {
+  final int categoryId;
+
+  const DeleteCategory(this.categoryId);
+
+  @override
+  List<Object?> get props => [categoryId];
+}